Language: Start working on the compiler by beginning the lexer
This patch actually starts making Tuploid a real language, rather than
just something I imagine.

This starts building towards a parser, but it only adds a basic lexer
at the moment. The plan is to get the lexer in a place where I'm
comfortable with it, and then work on parsing the language.

The parser is placed inside its own crate so that additional tooling
can be easily built on top of it in the future (for example an LSP
server).